Backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
awsebcli==3.23.1
bcrypt==4.2.1
blessed==1.20.0
botocore==1.37.37
cement==2.10.14
certifi==2024.8.30
charset-normalizer==3.4.1
cloudinary==1.41.0
colorama==0.4.6
Django==5.1.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
dnspython==2.6.1
ecdsa==0.19.0
fastapi==0.115.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
semantic-version==2.10.0
setuptools==79.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.38.6
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
